+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "symlink-follow.autoFollow": true
+}
