+++ modules/asset/ui/src/index.js
@@ -1,8 +1,4 @@
-import { parse } from './player.js';
 import './tailwind.css';
+import './main.scss';
 
-export default () => {
-  // Demonstrating how to pass helper functions to be used by other modules
-  // apart from a direct import via aliasing.
-  apos.util.parsePlayerData = parse;
-};
+export default () => {};
