--- modules/counter-page/ui/src/app/App.css
@@ -1,5 +1,0 @@
-.card {
-  padding: 2em;
-  text-align: center;
-}
-
