--- modules/counter-page/ui/src/app/index.css
@@ -1,24 +1,0 @@
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-.cbutton {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 2em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  cursor: pointer;
-  transition: border-color 0.25s;
-  color: white;
-}
-.cbutton:hover {
-  border-color: #646cff;
-}
-.cbutton:focus,
-.cbutton:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
