+++ modules/counter-react-widget/ui/src/counter-react.js
@@ -1,5 +1,6 @@
 import { createRoot } from 'react-dom/client';
 import { createElement } from 'react';
+import { parse } from '@/counter/src/player';
 import App from './app/App.jsx';
 
 // Environments are available here (`import.meta.env.PROD`, `import.meta.env.DEV`, etc.)
@@ -25,7 +26,7 @@ function player(el) {
     id,
     widget,
     options
-  } = apos.util.parsePlayerData(el);
+  } = parse(el);
   // 3. Mount and render the app
   const app = createElement(App, {
     id,
