+++ modules/counter-react-widget/views/widget.html
@@ -1,8 +1,10 @@
 {#
   Send the data to the counterApp component and let it render the markup.
   `options` is the object defined in the area schema. 
+  Use a generic `counterApp` component to render the markup.
+  See `modules/counter/views/counterApp.html`
  #} 
-{% component 'asset:counterApp' with {
+{% component 'counter:counterApp' with {
   framework: "react",
   widget: data.widget, 
   page: data.page,
