+++ modules/counter-vue-widget/ui/src/app/App.vue
@@ -12,7 +12,6 @@ const message = ref('');
 const debugState = ref(false);
 const debugLabel = computed(() => `${debugState.value ? 'Hide' : 'Show'} Debug`);
 const debug = JSON.stringify({ id, widget, options }, null, 2);
-const baseUrl = computed(() => window.apos.assetBaseUrl);
 
 const onClick = () => {
   message.value = "";
@@ -39,17 +38,17 @@ const onDebugClick = () => {
   <div class="py-8">
     <div class="flex justify-center content-center">
       <a href="https://vite.dev" target="_blank" rel="noreferrer">
-        <!-- We can use public images as we did -->
-        <img :src="`${baseUrl}/modules/asset/vite.svg`" class="logo" alt="Vite Logo" />
+        <!-- Use the alias and point to modules/asset/ui/svg/vite.svg -->
+        <img src="`@/asset/svg/vite.svg`" class="logo" alt="Vite Logo" />
       </a>
       <a href="https://vuejs.org/" target="_blank" rel="noreferrer">
-        <!-- ...or a relative path -->
+        <!-- ...or a relative path to the same module -->
         <img src="./assets/vue.svg" class="logo vue" alt="Vue Logo" />
       </a>
     </div>
 
     <!-- Title from the widget data  -->
-    <h1>{{ widget.title }}</h1>
+    <h2 class="text-5xl">{{ widget.title }}</h2>
 
     <!-- A server error message will appear here -->
     <p v-if="message" class="mt-4 p-4 bg-red-400">[Server Message] {{ message }}</p>
