+++ modules/counter/ui/src/index.js
@@ -1,0 +1,7 @@
+import { parse } from './player.js';
+
+export default () => {
+  // Demonstrating how one can pass helper functions to be used by other modules
+  // apart from a direct import via aliasing.
+  apos.util.parsePlayerData = parse;
+};
