+++ modules/counter/views/counterApp.html
@@ -3,3 +3,4 @@
   data-apos-widget-id="{{ data.widget._id }}" 
   data-apos-widget='{{ data.widget | merge({ counter: data.counter }) | toAttributeValue }}'
   data-apos-options='{{ data.options | toAttributeValue }}'></div>
+{# Async Server Component defined in `modules/counter/index.js` #}
