--- modules/default-page/index.js
@@ -1,59 +1,0 @@
-export default {
-  extend: '@apostrophecms/page-type',
-  options: {
-    label: 'Default Page'
-  },
-  fields: {
-    add: {
-      main: {
-        type: 'area',
-        options: {
-          widgets: {
-            '@apostrophecms/rich-text': {
-              toolbar: [
-                'styles',
-                '|',
-                'bold',
-                'italic',
-                'strike',
-                'link',
-                '|',
-                'bulletList',
-                'orderedList'
-              ],
-              styles: [
-                {
-                  tag: 'p',
-                  label: 'Paragraph (P)'
-                },
-                {
-                  tag: 'h3',
-                  label: 'Heading 3 (H3)'
-                },
-                {
-                  tag: 'h4',
-                  label: 'Heading 4 (H4)'
-                }
-              ],
-              insert: [
-                'table',
-                'image'
-              ]
-            },
-            '@apostrophecms/image': {},
-            '@apostrophecms/video': {}
-          }
-        }
-      }
-    },
-    group: {
-      basics: {
-        label: 'Basics',
-        fields: [
-          'title',
-          'main'
-        ]
-      }
-    }
-  }
-};
