+++ tailwind.config.js
@@ -1,7 +1,7 @@
 /** @type {import('tailwindcss').Config} */
 export default {
   content: [
-    './apos-build/@apostrophecms/vite/default/src/**/*.{js,jsx}',
+    './apos-build/@apostrophecms/vite/default/src/**/*',
     './modules/**/views/**/*.html',
     './views/**/*.html'
   ],
